Ya calcula soluciones factibles

+++ src/main/java/com/vrj/coh/tsp/LoteResponse.java
@@ -1,13 +1,15 @@
 package com.vrj.coh.tsp;
 
+import java.math.BigDecimal;
+
 import lombok.Data;
 
 @Data
 public class LoteResponse {
-    private double promedio;
+    private BigDecimal promedio;
     private Solution solution;
 
-    public LoteResponse(double promedio, Solution tsp) {
+    public LoteResponse(BigDecimal promedio, Solution tsp) {
         this.promedio = promedio;
         this.solution = tsp;
     }
